:memo: :art: Organize repo structure and add doc
**Problem**
The repository has a flat structure that make it difficult to understand
how to use it. Also there is no clear documentation of what the project
is doing.

**Solution**

* :art: Rename `data` folder to `cache` and add to repo
* Add screenshot of example dashboard output
* :art: Separate node modules for Github queries and cache
* :memo: Add readme with project description and todo list

**Result**
The repo is documented and node modules are better structured.
