+++ index.js
@@ -1,0 +1,13 @@
+var github = require('./repo-loader.js')
+var github = require('./repo-metrics.js')
+var Q = require('q')
+var _ = require('lodash')
+
+/**
+* @module Get Github organizations metrics
+* @author Camille Mizony
+* @TODO write interface prototypes
+* @version 1.0.0
+*/
+module.exports = {
+}
